backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
django-storages==1.14.2
boto3==1.29.7
pytest==7.4.3
pytest-django==4.7.0
hypothesis==6.88.1
factory-boy==3.3.0
coverage==7.3.2
black==23.10.1
flake8==6.1.0
pre-commit==3.5.0